requirements.txt
# Flask y extensiones
Flask==3.0.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0

# Seguridad y autenticación
passlib==1.7.4
bcrypt==4.1.2
itsdangerous==2.1.2
authlib==1.3.0

# PDF y generación de documentos
reportlab==4.1.0
pdfkit==1.0.0

# Archivos y formularios
python-dotenv==1.0.1
Werkzeug==3.0.2

# Validación y datos
email-validator==2.1.1
validators==0.28.0
marshmallow==3.21.1
python-dateutil==2.9.0.post0
pytz==2024.1

# APIs y conexiones externas
requests==2.32.3
httpx==0.27.0

# Webcam y multimedia (opcional, prueba opencv-python-headless)
opencv-python-headless==4.9.0.80
imageio==2.34.1

# Archivos masivos y Excel
openpyxl==3.1.2
xlrd==2.0.1

# Gráficos y visualización
seaborn==0.13.2
plotly==5.22.0

# Desarrollo y pruebas
gunicorn==22.0.0
ipython==8.24.0
pytest==8.2.1
coverage==7.5.1
flake8==7.0.0
